refactor: migrate classproperty to separate file
