+++ discatcore/utils/event.py
@@ -5,27 +5,15 @@ from __future__ import annotations
 import typing as t
 
 import attr
-from typing_extensions import Self
+
+from .functools import classproperty
 
 if t.TYPE_CHECKING:
     from .dispatcher import ListenerCallback
 
 __all__ = ("Event", "ExceptionEvent")
 
-T = t.TypeVar("T")
 EventT = t.TypeVar("EventT", bound="Event")
-
-
-class _classproperty(t.Generic[T]):
-    def __init__(self, fget: t.Callable[[t.Any], T], /) -> None:
-        self.fget: "classmethod[T]" = t.cast("classmethod[T]", fget)
-
-    def getter(self, fget: t.Callable[[t.Any], T], /) -> Self:
-        self.fget = t.cast("classmethod[T]", fget)
-        return self
-
-    def __get__(self, obj: t.Optional[t.Any], type: t.Optional[type]) -> T:
-        return self.fget.__func__(type)
 
 
 class Event:
@@ -40,7 +28,7 @@ class Event:
 
         cls.__dispatches = tuple(base for base in cls.__mro__ if issubclass(base, Event))
 
-    @_classproperty
+    @classproperty
     @classmethod
     def dispatches(cls):
         return cls.__dispatches
